\name{ji.ds.heatmapplot}
\alias{ji.ds.heatmapplot}
\title{Generates a heatmap plot for merged datasets}
\usage{
  ji.ds.heatmapplot(opals, xvect, yvect)
}
\arguments{
  \item{opals}{character strings that represent the URL of
  the servers where the study datasets are stored.}

  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a heatmap plot
}
\description{
  Generates a heatmap plot for merged datasets
}

